requirements.txt
pyusb==0.4.3
pygobject==3.12.0
requests==2.2.1
